Add files via upload
Added the mod file for Mundane Zombies. I updated the original obsoleted mod to include the new monsters that weren't already in. They are the zombie biter, zombie mancroc, and the zombie brute ninja and zombie brute grappler. I did not add in the variants of the zombie child so things like sproglodytes and snotgobblers will still exist. If someone feels these need to be removed as well for the sake of thoroughness, let me know and I will update the mod appropriately.

Smoky Bear was removed from the exclusion list because there are no other zombie animals listed in the mod, and if this was intended to be one of the joke monsters [literally an undead "smoky the bear" reference] then it would be more appropriately placed in the Crazy Cataclysm mod.

Mundane Zombies now focuses exclusively on removing any mutated humanoid zombies, leaving only the John Romero-style [but World War Z speed] shamblers.